+++ inc/classes/bs-user.php
@@ -3,16 +3,23 @@
 class BS_User {
     private $user_details = array(
         'id' => 0,
-        'name' => 'Anon',
         'email' => 'anon@nymous.org',
         'password' => 'password',
         'level' => 0,
+        'logged_in' => 0
     );
 
-    public function __construct( $user_details = false ) {
-        if ( is_array( $user_details ) ) {
-            return $this->load( $user_details );
+    public function __construct( $mode = '' ) {
+        switch ( $mode ) {
+        case 'current':
+            // Try to load from session
+            if ( isset( $_SESSION['bs_user'] ) && !empty( $_SESSION['bs_user'] ) ) {
+                $this->user_details = $_SESSION['bs_user'];
+            }
+            return $this->user_details;
+            break;
         }
+        return false;
     }
 
     public function create( $db = false, $user_details = array() ) {
@@ -49,7 +56,7 @@ class BS_User {
             return $errors;
         }
         else {
-            $user_exists = $db->select_where( 'user', array(
+            $user_exists = $this->get_user_by( $db, array(
                     'email' => $values['email']
                 ) );
             // Create user
@@ -62,15 +69,36 @@ class BS_User {
         }
     }
 
-    public function load( $user_details = array() ) {
-
+    private function get_user_by( $db, $details ) {
+        foreach ( $details as $id => $var ) {
+            if ( $id == 'password' ) {
+                $details[$id] = $this->hashPassword( $var );
+            }
+        }
+        return $db->select_where( 'user', $details );
     }
 
-    public function validateDetails( $user_details = array() ) {
-
+    public function connect( $db, $details ) {
+        $user = $this->get_user_by( $db, $details );
+        if ( empty( $user ) ) {
+            return false;
+        }
+        return $this->login( $user[0] );
     }
 
-    public function invalidValue( $id, $user_details ) {
+    public function disconnect() {
+        $_SESSION['bs_user'] = array();
+    }
+
+    private function login( $user ) {
+        $user['logged_in'] = 1;
+        $_SESSION['bs_user'] = $user;
+        $this->user_details = $user;
+        return true;
+    }
+
+
+    private function invalidValue( $id, $user_details ) {
         $test = true;
         switch ( $id ) {
         case 'email':
@@ -88,9 +116,23 @@ class BS_User {
         return $test;
     }
 
-    public function hashPassword( $password ) {
+    private function hashPassword( $password ) {
         // _todo: improve this <3
         $hash = md5( '?salt?' . $password );
         return $hash;
+    }
+
+    // Getters
+    public function getEmail() {
+        return $this->user_details['email'];
+    }
+    public function getID() {
+        return $this->user_details['id'];
+    }
+    public function getLevel() {
+        return $this->user_details['level'];
+    }
+    public function isLoggedIn() {
+        return $this->user_details['logged_in'];
     }
 }